build resources with webpack
