+++ package.json
@@ -1,0 +1,91 @@
+{
+    "name": "collective-z3cform-datagridfield",
+    "version": "1.0.0",
+    "description": "DatagridField Widget for z3c.form",
+    "license" : "BSD-3-Clause",
+    "engines": {
+        "node": ">=8.12.0",
+        "yarn": "^1.3.2"
+    },
+    "dependencies": {
+        "@patternslib/patternslib": "^7.4.0",
+        "mockup": "git+https://github.com/plone/mockup.git#master"
+    },
+    "devDependencies": {
+        "@babel/core": "^7.16.0",
+        "@babel/eslint-parser": "^7.16.3",
+        "@babel/plugin-proposal-optional-chaining": "^7.16.0",
+        "@babel/preset-env": "^7.16.0",
+        "@testing-library/jest-dom": "^5.15.0",
+        "@types/sinon": "^10.0.2",
+        "babel-loader": "^8.2.3",
+        "clean-css-cli": "^4.3.0",
+        "copy-webpack-plugin": "^10.0.0",
+        "copyfiles": "^2.4.1",
+        "core-js": "3.20.2",
+        "cross-env": "^7.0.3",
+        "css-loader": "^6.5.0",
+        "eslint": "^8.2.0",
+        "eslint-config-prettier": "^8.3.0",
+        "expose-loader": "^3.0.0",
+        "identity-obj-proxy": "^3.0.0",
+        "imports-loader": "^3.1.0",
+        "inspectpack": "^4.7.1",
+        "jest": "^27.3.1",
+        "jest-raw-loader": "^1.0.1",
+        "jest-watch-typeahead": "^1.0.0",
+        "npm-run-all": "^4.1.5",
+        "nodemon": "^2.0.15",
+        "prettier": "^2.4.1",
+        "prism-themes": "^1.9.0",
+        "regenerator-runtime": "^0.13.9",
+        "sass": "^1.43.4",
+        "sass-loader": "^12.3.0",
+        "select2": "^4.1.0",
+        "style-loader": "^3.3.0",
+        "terser-webpack-plugin": "^5.2.5",
+        "timezone-mock": "^1.2.2",
+        "webpack": "^5.64.1",
+        "webpack-bundle-analyzer": "^4.4.2",
+        "webpack-cli": "^4.9.1",
+        "webpack-dev-server": "^4.5.0",
+        "webpack-modernizr-loader": "^5.0.0",
+        "whybundled": "^2.0.0",
+        "yarn": "^1.22.17"
+    },
+    "resolutions": {
+        "jquery": "3.6.0"
+    },
+    "resolutions-comments": {
+        "ua-parser-js": "See https://github.com/faisalman/ua-parser-js/issues/536"
+    },
+    "scripts": {
+        "build": "npm-run-all build:webpack build:css",
+        "build:webpack": "NODE_ENV=production webpack --config webpack.config.js",
+        "build:css": "npm-run-all css-compile css-minify",
+        "watch:webpack": "NODE_ENV=development webpack --config webpack.config.js --watch",
+        "watch:scss": "nodemon --watch resources/scss/ --ext scss --exec \"yarn build:css\"",
+        "start": "NODE_ENV=development webpack serve --config webpack.config.js",
+        "css-compile": "sass --load-path=node_modules --style expanded --source-map --embed-sources --no-error-css resources/scss/datagridfield.scss:src/collective/z3cform/datagridfield/static/datagridfield.css",
+        "css-minify": "cleancss --level 1 --format breakWith=lf --source-map --source-map-inline-sources --output src/collective/z3cform/datagridfield/static/datagridfield.min.css src/collective/z3cform/datagridfield/static/datagridfield.css"
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ]
+    },
+    "publishConfig": {
+        "access": "public"
+    },
+    "author": {
+        "name": "petschki",
+        "email": "peter.mathis@kombinat.at"
+    },
+    "homepage": "https://github.com/collective/collective.z3cform.datagridfield",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/collective/collective.z3cform.datagridfield.git"
+    }
+}
