+++ resources/datagridfield-config.js
@@ -1,0 +1,4 @@
+import registry from "@patternslib/patternslib/src/core/registry";
+import "./js/datagridfield.pattern";
+
+registry.init();
