--- src/collective/z3cform/datagridfield/static/datagridfield-bundle-compiled.css
@@ -1,1 +1,0 @@
-.datagridwidget-table-view{width:100%;border:1px solid #666;background:none repeat scroll 0 0 #eee}.datagridwidget-table-view .header{font-weight:bold;background:none repeat scroll 0 0 #eee;color:#666;padding:2px;font-size:1.1em;text-align:left}.datagridwidget-row{vertical-align:top}.datagridwidget-cell{vertical-align:bottom;padding:1px}.datagridwidget-cell input[type="text"]{padding:1px}.datagridwidget-cell input{border:1px solid #ccc;background:white;background-image:none}.datagridwidget-cell select{border:1px solid #ccc}.datagridwidget-cell textarea{border:1px solid #ccc}.datagridwidget-cell>input:only-child{width:99%}.datagridwidget-cell div.error{color:red;font-weight:bold}.datagridwidget-empty-row{display:none}.datagridwidget-manipulator{width:16px;text-align:center;padding-bottom:4px;margin:0;vertical-align:bottom}.datagridwidget-manipulator img{cursor:pointer}.datagridwidget-hidden-data{display:none}.datagridwidget-table-view-block{border-collapse:collapse}.datagridwidget-table-view-block thead{display:none}.datagridwidget-table-view-block .datagridwidget-row{border-bottom:1px solid #eee}.datagridwidget-table-view-block td{vertical-align:top}.datagridwidget-table-view-block .datagridwidget-manipulator{width:16px}.datagridwidget-block-edit-cell{width:auto}.datagridwidget-block-edit-cell .error{background-color:#ffddcc;border:1px solid red;margin:1em 0;padding:.5em}.datagridwidget-block{padding:.5em}/*# sourceMappingURL=datagridfield-bundle-compiled.css.map */
