+++ src/collective/z3cform/datagridfield/static/datagridfield.css
@@ -1,140 +1,94 @@
-/*
-  Attempted to keep the CSS similar to the Product.DataGridWidget definitioans.
-
-  IE Hacks from that version or left in on the grounds that they might do something,
-*/
-
 .datagridwidget-table-view {
-	width: 100%;
-    border: 1px solid #666;
-    background: none repeat scroll 0 0 #EEE;
+  width: 100%;
+  border: 1px solid #666;
+  background: none repeat scroll 0 0 #eee;
 }
-
 .datagridwidget-table-view .header {
-    font-weight: bold;
-    background: none repeat scroll 0 0 #DDD;
-    color: #666;
-    padding: 2px;
-    font-size: 1.1em;
-    text-align: left;
+  font-weight: bold;
+  background: none repeat scroll 0 0 #eee;
+  color: #666;
+  padding: 2px;
+  font-size: 1.1em;
+  text-align: left;
 }
 
 .datagridwidget-row {
-	vertical-align: top;
+  vertical-align: top;
 }
 
-.datagridwidget-cell input[type="text"] {
-    padding: 1px;
+.datagridwidget-cell {
+  vertical-align: bottom;
+  padding: 1px;
 }
-.datagridwidget-cell input, .datagridwidget-cell select, .datagridwidget-cell textarea {
-    border: 1px solid #CCC;
+.datagridwidget-cell input[type=text] {
+  padding: 1px;
 }
-.datagridwidget-cell > input:only-child { width: 99%; }
-
+.datagridwidget-cell input {
+  border: 1px solid #ccc;
+  background: white;
+  background-image: none;
+}
+.datagridwidget-cell select {
+  border: 1px solid #ccc;
+}
+.datagridwidget-cell textarea {
+  border: 1px solid #ccc;
+}
+.datagridwidget-cell > input:only-child {
+  width: 99%;
+}
 .datagridwidget-cell div.error {
-    color:red;
-    font-weight:bold;
+  color: red;
+  font-weight: bold;
 }
 
 .datagridwidget-empty-row {
-	/* Hidden row with default selection widget values */
-	display: none;
-}
-
-/* Error message is displayed above the field */
-.datagridwidget-cell {
-	vertical-align: bottom;
-	padding: 1px;
+  display: none;
 }
 
 .datagridwidget-manipulator {
-	width: 16px;
-	text-align: center;
-	padding-bottom: 4px;
-	margin: 0;
-	vertical-align: bottom;
+  width: 16px;
+  text-align: center;
+  padding-bottom: 4px;
+  margin: 0;
+  vertical-align: bottom;
 }
-
 .datagridwidget-manipulator img {
-	cursor: pointer;
+  cursor: pointer;
 }
 
 .datagridwidget-hidden-data {
-	display: none;
+  display: none;
 }
-
-/* You cannot delete the auto-append row */
-.auto-append  > .datagridwidget-manipulator.delete-row a,
-.auto-append  > .datagridwidget-manipulator.move-up a,
-.auto-append  > .datagridwidget-manipulator.move-down a,
-.auto-append  > .datagridwidget-manipulator.insert-row a {
-    display: none;
-}
-
-/* When no AA mode, only one row in the widget */
-.minimum-row  > .datagridwidget-manipulator.insert-row a {
-    display: block;
-}
-
-.datagridwidget-body-non-auto-append[data-many-rows = "true"] .datagridfield-last-filled-row > .datagridwidget-manipulator.move-down a,
-.datagridwidget-body-non-auto-append[data-many-rows = "true"] .datagridfield-first-filled-row > .datagridwidget-manipulator.move-up a {
-    display: block;
-}
-
-/* Cannot move up out of the grid */
-.datagridfield-first-filled-row > .datagridwidget-manipulator.move-up a {
-    display: none !important;
-}
-
-/* Cannot move below auto-append row */
-.datagridfield-last-filled-row > .datagridwidget-manipulator.move-down a {
-    display: none !important;
-}
-
-/*
- 	Workaround for IE which cannot propeply handle
- 	background images in cloned TR input fields.
- 	Plone is shipped with default input background gradient image
- */
-.datagridwidget-cell input {
-	background: white;
-	background-image: none;
-}
-
-/**
- * Block edit mode styles
- */
 
 .datagridwidget-table-view-block {
-    border-collapse: collapse;
+  border-collapse: collapse;
 }
-
 .datagridwidget-table-view-block thead {
-    display: none;
+  display: none;
 }
-
 .datagridwidget-table-view-block .datagridwidget-row {
-    border-bottom: 1px solid black;
+  border-bottom: 1px solid #eee;
 }
-
 .datagridwidget-table-view-block td {
-    vertical-align: top;
+  vertical-align: top;
+}
+.datagridwidget-table-view-block .datagridwidget-manipulator {
+  width: 16px;
 }
 
 .datagridwidget-block-edit-cell {
-    width: auto;
+  width: auto;
 }
-.datagridwidget-table-view-block .datagridwidget-manipulator {
-    width: 16px;
+.datagridwidget-block-edit-cell .error {
+  background-color: #ffddcc;
+  border: 1px solid red;
+  margin: 1em 0;
+  padding: 0.5em;
 }
 
 .datagridwidget-block {
-    padding: 0.5em;
+  padding: 0.5em;
 }
 
-.datagridwidget-block-edit-cell .error {
-    background-color: #FFDDCC;
-    border: 1px solid red;
-    margin: 1em 0;
-    padding: 0.5em;
-}
+/*# sourceMappingURL=datagridfield.css.map */
