+++ src/collective/z3cform/datagridfield/static/polyfills-loader.js
@@ -1,0 +1,1 @@
+var script_source=function(){var t=document.getElementsByTagName("script"),e=t[t.length-1];return void 0!==e.getAttribute.length?e.src:e.getAttribute("src",-1)};!function(){if(document.documentMode){var t=script_source();t=t.substring(0,t.lastIndexOf("/"))+"/";var e=document.createElement("script");e.type="text/javascript",e.src=t+"bundle-polyfills.min.js",document.getElementsByTagName("head")[0].appendChild(e)}}();
